Moving parseHeader
