+++ lib/connection/connection.js
@@ -7,7 +7,7 @@ var inherits = require('util').inherits
   , crypto = require('crypto')
   , f = require('util').format
   , debugOptions = require('./utils').debugOptions
-  , parseHeader = require('./utils').parseHeader
+  , parseHeader = require('../wireprotocol/shared').parseHeader
   , compressorIDs = require('./utils').compressorIDs
   , Response = require('./commands').Response
   , MongoError = require('../error')
