+++ lib/connection/utils.js
@@ -97,16 +97,6 @@ var retrieveSnappy = function() {
   return snappy;
 }
 
-// Parses the header of a wire protocol message
-var parseHeader = function(message) {
-  return {
-    length: message.readInt32LE(0),
-    requestId: message.readInt32LE(4),
-    responseTo: message.readInt32LE(8),
-    opCode: message.readInt32LE(12)
-  }
-}
-
 exports.compressorIDs = {
   snappy: 1,
   zlib: 2
@@ -118,4 +108,3 @@ exports.copy = copy;
 exports.debugOptions = debugOptions;
 exports.retrieveBSON = retrieveBSON;
 exports.retrieveSnappy = retrieveSnappy;
-exports.parseHeader = parseHeader;
