+++ lib/wireprotocol/shared.js
@@ -37,8 +37,19 @@ var getReadPreference = function(cmd, options) {
   return readPreference;
 }
 
+// Parses the header of a wire protocol message
+var parseHeader = function(message) {
+  return {
+    length: message.readInt32LE(0),
+    requestId: message.readInt32LE(4),
+    responseTo: message.readInt32LE(8),
+    opCode: message.readInt32LE(12)
+  }
+}
+
 module.exports = {
   getReadPreference: getReadPreference,
   MESSAGE_HEADER_SIZE: MESSAGE_HEADER_SIZE,
-  opcodes: opcodes
+  opcodes: opcodes,
+  parseHeader: parseHeader
 }
